fix observation in the code review
*I think you can separate this class in two classes.
ResourcesSteps
ResponsesSteps

*This variable should be attribute of class.
In order to re-use it.

*  When I did a post  with:
      | Content    | Icon |
      | Testeo     | 4     |
